backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
beautifulsoup4==4.12.3
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
pydantic==2.5.3
httpx==0.26.0
lxml==5.1.0
groq==0.4.1